完成了登录功能

+++ src/main/java/com/example/mybatisplus/model/dto/UserInfoDTO.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class UserInfoDTO {
     private Long id;
     private String name;
-    private Long userType;
+    private Integer userType;
 }
